Moved Login.js to src and change LoginPage.js import to reflect change
